Clear form on submit and remove delete button
When submitting a new gift the 'new gift' form should be cleared. This
proved a little difficult since the form consist completely of
uncontrolled fields. I now assign a key to that form. When a new gift is
submitted I increment the key which renders a new (and thus empty) form.

I also removed the 'delete' button on the 'new gift' form since it had
no place there and would just error out.